1.Template活用、2.3層アーキテクチャ、3登録済の場合はダッシュボード遷移
